+++ R/read1d.R
@@ -18,12 +18,17 @@
 #' @author Torben Kimhofer \email{torben.kimhofer@@murdoch.edu.au}
 #' @family NMR
 #' @importFrom stats approxfun
+#' @importFrom plyr ddply
 #' @seealso \code{\link[=read1d_raw]{Process raw FIDs}}
 #' @section
 
 read1d <- function(path, exp_type = list(exp = c("PROF_PLASMA_CPMG")),
     n_max = 1000, filter = TRUE, recursive = TRUE, verbose = TRUE) {
     path <- path.expand(path)
+
+    if (as.character(match.call()[[1]]) == "read1d") {
+        warning("read1d` will be removed in future versions, please use `read1d_proc` instead.", call. = FALSE)
+    }
 
     f_list <- .detect1d_procs(path, n_max = 1e6, filter, recursive, verbose)
     # extract parameters from acqus and procs (for f1 and f2)
@@ -41,7 +46,7 @@ read1d <- function(path, exp_type = list(exp = c("PROF_PLASMA_CPMG")),
     # !any(is.na(suppressWarnings(as.numeric(x)))) }) pars<-as.data.frame(pars)
     # pars[,dtype_num]<-apply(pars[,dtype_num], 2, as.numeric)
     # rownames(pars)<-f_list[[2]]
-    # chem shift
+        # chem shift
     ppm_ref <- .chemShift(swidth = pars$a_SW[1], offset = pars$p_OFFSET[1], si = pars$p_SI[1])
     # read binary file (=spectrum)
     out <- vapply(seq_along(f_list[[1]]), function(s, ppref = ppm_ref) {
@@ -75,7 +80,9 @@ read1d <- function(path, exp_type = list(exp = c("PROF_PLASMA_CPMG")),
     assign("meta", pars, envir = .GlobalEnv)
 }
 
-
+#' @export
+#' @rdname read1d
+read1d_proc <- read1d
 
 
 #' @title Read Bruker NMR paramter files - helper function read1d