+++ NEWS.md
@@ -1,3 +1,11 @@
-# metabom8 0.1.0
+# metabom8 0.4.4
 
-* Added a `NEWS.md` file to track changes to the package.
+* Added `alignSegment()` for aligning spectral regions with peak shift
+* Updated NIPALS convergence threhold for PCA (eps<10^-6^)
+
+
+# metabom8 0.4.3
+
+* Updated `binning()` function
+* Added `excise1d()` for automated 1D chemical shift excision (capping, water, urea)
+* Added `normErectic()` to normalise std 1D spec with the ERECTIC integral (Bruker IVDr method)